update to decimal

+++ lending-engine/app/api/serializer.py
@@ -106,7 +106,7 @@ class BniRdlCallbackSchema(ma.Schema):
     """ this is schema for callback object """
     p2p_id = fields.Str(required=True, validate=cannot_be_blank)
     account_number = fields.Str(required=True, validate=cannot_be_blank)
-    payment_amount = fields.Float(required=True, validate=cannot_be_blank)
+    payment_amount = fields.Decimal(required=True, validate=cannot_be_blank)
     accounting_flag = fields.Str(required=True, validate=cannot_be_blank)
     journal_number = fields.Int(required=True, validate=cannot_be_blank)
     datetime_payment = fields.Str(required=True, validate=cannot_be_blank)
